Adjusting logic on re-registering the Node in mDNC case
