+++ cns/NetworkContainerContract.go
@@ -373,8 +373,7 @@ func (networkContainerRequestPolicy *NetworkContainerRequestPolicies) Validate()
 
 // NodeInfoResponse - Struct to hold the node info response.
 type NodeInfoResponse struct {
-	NetworkContainers  []CreateNetworkContainerRequest
-	NmAgentApisMissing bool
+	NetworkContainers []CreateNetworkContainerRequest
 }
 
 // NodeRegisterRequest - Struct to hold the node register request.
