+++ cns/restserver/internalapi.go
@@ -10,9 +10,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"runtime"
-	"strconv"
-	"time"
 
 	"github.com/Azure/azure-container-networking/cns"
 	"github.com/Azure/azure-container-networking/cns/logger"
@@ -126,14 +123,6 @@ func (service *HTTPRestService) SyncNodeStatus(dncEP, infraVnet, nodeID string, 
 	service.Lock()
 	service.saveState()
 	service.Unlock()
-
-	if nodeInfoResponse.NmAgentApisMissing {
-		// RegisterNode again with NmAgent Apis list
-		retErr := RegisterNode(httpc, service, dncEP, infraVnet, nodeID)
-		if retErr != nil {
-			logger.Errorf("[Azure-CNS] Failed to register Node ID: %s with error: %s", nodeID, err.Error())
-		}
-	}
 
 	// delete dangling NCs
 	for nc := range ncsToBeDeleted {
@@ -261,90 +250,4 @@ func (service *HTTPRestService) CreateOrUpdateNetworkContainerInternal(req cns.C
 	}
 
 	return returnCode
-}
-
-// RegisterNode - Tries to register node with DNC when CNS is started in managed DNC mode
-func RegisterNode(httpc *http.Client, httpRestService cns.HTTPService, dncEP, infraVnet, nodeID string) error {
-	logger.Printf("[Azure CNS] Registering node %s with Infrastructure Network: %s PrivateEndpoint: %s", nodeID, infraVnet, dncEP)
-
-	var (
-		numCPU              = runtime.NumCPU()
-		url                 = fmt.Sprintf(common.RegisterNodeURLFmt, dncEP, infraVnet, nodeID, dncApiVersion)
-		nodeRegisterRequest cns.NodeRegisterRequest
-	)
-
-	nodeRegisterRequest.NumCPU = numCPU
-	supportedApis, retErr := nmagentclient.GetNmAgentSupportedApis(httpc, "")
-
-	if retErr != nil {
-		logger.Errorf("[Azure CNS] Failed to retrieve SupportedApis from NMagent of node %s with Infrastructure Network: %s PrivateEndpoint: %s",
-			nodeID, infraVnet, dncEP)
-		return retErr
-	}
-
-	//To avoid any null-pointer deferencing errors.
-	if supportedApis == nil {
-		supportedApis = []string{}
-	}
-
-	nodeRegisterRequest.NmAgentSupportedApis = supportedApis
-
-	nodeRegisterTicker := time.NewTicker(common.FiveSeconds)
-	responseChan := make(chan error)
-
-	for {
-		select {
-		case responseErr := <-responseChan:
-			return responseErr
-		case <-nodeRegisterTicker.C:
-			go sendRegisterNodeRequest(httpc, httpRestService, nodeRegisterRequest, url, responseChan)
-		}
-	}
-}
-
-// sendRegisterNodeRequest func helps in registering the node until there is an error.
-func sendRegisterNodeRequest(
-	httpc *http.Client,
-	httpRestService cns.HTTPService,
-	nodeRegisterRequest cns.NodeRegisterRequest,
-	registerURL string,
-	responseChan chan<- error) {
-
-	var (
-		body     bytes.Buffer
-		response *http.Response
-		err      = fmt.Errorf("")
-	)
-
-	err = json.NewEncoder(&body).Encode(nodeRegisterRequest)
-	if err != nil {
-		log.Errorf("encoding json failed with %v", err)
-		responseChan <- err
-		return
-	}
-
-	response, err = httpc.Post(registerURL, "application/json", &body)
-	if err == nil {
-		if response.StatusCode == http.StatusCreated {
-			var req cns.SetOrchestratorTypeRequest
-			decodeErr := json.NewDecoder(response.Body).Decode(&req)
-			if decodeErr != nil {
-				log.Errorf("decoding Node Resgister response json failed with %v", err)
-				responseChan <- err
-				return
-			}
-			httpRestService.SetNodeOrchestrator(&req)
-
-			logger.Printf("[Azure CNS] Node Registered")
-			responseChan <- nil
-		} else {
-			err = fmt.Errorf("[Azure CNS] Failed to register node with http status code %s", strconv.Itoa(response.StatusCode))
-			logger.Errorf(err.Error())
-			responseChan <- err
-		}
-
-		response.Body.Close()
-	} else {
-		logger.Errorf("[Azure CNS] Failed to register node with err: %+v", err)
-	}
 }
