+++ cns/service/main.go
@@ -4,11 +4,15 @@
 package main
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"os/signal"
+	"runtime"
+	"strconv"
 	"strings"
 	"syscall"
 	"time"
@@ -16,6 +20,7 @@ import (
 	localtls "github.com/Azure/azure-container-networking/server/tls"
 
 	"github.com/Azure/azure-container-networking/cns/ipampoolmonitor"
+	"github.com/Azure/azure-container-networking/cns/nmagentclient"
 
 	"github.com/Azure/azure-container-networking/aitelemetry"
 	"github.com/Azure/azure-container-networking/cnm/ipam"
@@ -243,6 +248,92 @@ func printVersion() {
 	fmt.Printf("Version %v\n", version)
 }
 
+// RegisterNode - Tries to register node with DNC when CNS is started in managed DNC mode
+func registerNode(httpc *http.Client, httpRestService cns.HTTPService, dncEP, infraVnet, nodeID string) error {
+	logger.Printf("[Azure CNS] Registering node %s with Infrastructure Network: %s PrivateEndpoint: %s", nodeID, infraVnet, dncEP)
+
+	var (
+		numCPU              = runtime.NumCPU()
+		url                 = fmt.Sprintf(acn.RegisterNodeURLFmt, dncEP, infraVnet, nodeID, dncApiVersion)
+		nodeRegisterRequest cns.NodeRegisterRequest
+	)
+
+	nodeRegisterRequest.NumCPU = numCPU
+	supportedApis, retErr := nmagentclient.GetNmAgentSupportedApis(httpc, "")
+
+	if retErr != nil {
+		logger.Errorf("[Azure CNS] Failed to retrieve SupportedApis from NMagent of node %s with Infrastructure Network: %s PrivateEndpoint: %s",
+			nodeID, infraVnet, dncEP)
+		return retErr
+	}
+
+	//To avoid any null-pointer deferencing errors.
+	if supportedApis == nil {
+		supportedApis = []string{}
+	}
+
+	nodeRegisterRequest.NmAgentSupportedApis = supportedApis
+
+	nodeRegisterTicker := time.NewTicker(acn.FiveSeconds)
+	responseChan := make(chan error)
+
+	for {
+		select {
+		case responseErr := <-responseChan:
+			return responseErr
+		case <-nodeRegisterTicker.C:
+			go sendRegisterNodeRequest(httpc, httpRestService, nodeRegisterRequest, url, responseChan)
+		}
+	}
+}
+
+// sendRegisterNodeRequest func helps in registering the node until there is an error.
+func sendRegisterNodeRequest(
+	httpc *http.Client,
+	httpRestService cns.HTTPService,
+	nodeRegisterRequest cns.NodeRegisterRequest,
+	registerURL string,
+	responseChan chan<- error) {
+
+	var (
+		body     bytes.Buffer
+		response *http.Response
+		err      = fmt.Errorf("")
+	)
+
+	err = json.NewEncoder(&body).Encode(nodeRegisterRequest)
+	if err != nil {
+		log.Errorf("encoding json failed with %v", err)
+		responseChan <- err
+		return
+	}
+
+	response, err = httpc.Post(registerURL, "application/json", &body)
+	if err == nil {
+		if response.StatusCode == http.StatusCreated {
+			var req cns.SetOrchestratorTypeRequest
+			decodeErr := json.NewDecoder(response.Body).Decode(&req)
+			if decodeErr != nil {
+				log.Errorf("decoding Node Resgister response json failed with %v", err)
+				responseChan <- err
+				return
+			}
+			httpRestService.SetNodeOrchestrator(&req)
+
+			logger.Printf("[Azure CNS] Node Registered")
+			responseChan <- nil
+		} else {
+			err = fmt.Errorf("[Azure CNS] Failed to register node with http status code %s", strconv.Itoa(response.StatusCode))
+			logger.Errorf(err.Error())
+			responseChan <- err
+		}
+
+		response.Body.Close()
+	} else {
+		logger.Errorf("[Azure CNS] Failed to register node with err: %+v", err)
+	}
+}
+
 // Main is the entry point for CNS.
 func main() {
 	// Initialize and parse command line arguments.
@@ -418,7 +509,7 @@ func main() {
 		httpRestService.SetOption(acn.OptInfrastructureNetworkID, infravnet)
 		httpRestService.SetOption(acn.OptNodeID, nodeID)
 
-		registerErr := restserver.RegisterNode(acn.GetHttpClient(), httpRestService, privateEndpoint, infravnet, nodeID)
+		registerErr := registerNode(acn.GetHttpClient(), httpRestService, privateEndpoint, infravnet, nodeID)
 		if registerErr != nil {
 			logger.Errorf("[Azure CNS] Resgistering Node failed with error: %v PrivateEndpoint: %s InfrastructureNetworkID: %s NodeID: %s",
 				registerErr,
